requirements.txt
python-telegram-bot==20.0
gspread==5.0.0
oauth2client==4.1.3
python-dotenv==0.19.2